// src/generator/dmmf/types.ts
export type ReadonlyDeep<O> = {
  +readonly [K in keyof O]: ReadonlyDeep<O[K]>;
};

export namespace DMMF {
// ...
  export type Query = ReadonlyDeep<{
    name: string;
    args: SchemaArg[];
    output: QueryOutput;
  }>;
// ...
  export type OutputSchemaField = SchemaField &
    ReadonlyDeep<{
      argsTypeName: string | undefined;
    }>;
// ...
  export type ModelMapping = ReadonlyDeep<{
    // model: string;
    // plural: string;
    // findUnique?: string | null;
    // findUniqueOrThrow?: string | null;
    // findFirst?: string | null;
    // findFirstOrThrow?: string | null;
    // findMany?: string | null;
    // createOne?: string | null;
    // createMany?: string | null;
    // createManyAndReturn?: string | null;
    // updateOne?: string | null;
    // updateMany?: string | null;
    // upsertOne?: string | null;
    // deleteOne?: string | null;
    // deleteMany?: string | null;
    // aggregate?: string | null;
    // groupBy?: string | null;
    // count?: string | null;
    // findRaw?: string | null;
    // aggregateRaw?: string | null;

    // additional props
    modelName: string;
    actions: Action[];
    collectionName: string;
    resolverName: string;
    modelTypeName: string;
  }>;
  export enum ModelAction {
    findUnique = "findUnique",
    findUniqueOrThrow = "findUniqueOrThrow",
    findFirst = "findFirst",
    findFirstOrThrow = "findFirstOrThrow",
    findMany = "findMany",
    createOne = "createOne",
    createMany = "createMany",
    createManyAndReturn = "createManyAndReturn",
    updateOne = "updateOne",
    updateMany = "updateMany",
    upsertOne = "upsertOne",
    deleteOne = "deleteOne",
    deleteMany = "deleteMany",
    groupBy = "groupBy",
    // count = "count",
    aggregate = "aggregate",
    findRaw = "findRaw",
    aggregateRaw = "aggregateRaw",
  }
  // additional type
  export type Action = ReadonlyDeep<{
    name: string;
    fieldName: string;
    kind: ModelAction;
    operation: "Query" | "Mutation";
    prismaMethod: string;
    method: OutputSchemaField;
    argsTypeName: string | undefined;
    outputTypeName: string;
    actionResolverName: string;
    returnTSType: string;
    typeGraphQLType: string;
  }>;
  // additional type
  export type RelationModel = ReadonlyDeep<{
    model: Model;
    outputType: OutputType;
    relationFields: RelationField[];
    resolverName: string;
  }>;
  // additional type
  export type RelationField = ModelField &
    ReadonlyDeep<{
      outputTypeField: OutputSchemaField;
      argsTypeName: string | undefined;
    }>;
}